+++ app/views/gallery.js
@@ -1,0 +1,5 @@
+document.addEventListener('keydown', (event) => {
+    if (event.keyCode === 27) {
+        $('.uk-modal.uk-open').removeClass('uk-open').attr('aria-hidden', true).css('display', 'none');
+    }
+});