+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 ## 0.3.0
 - added image caching
 - added function to rotate images
+- added Escape Key listener to close lightbox
+- bugfixes
 
 ## 0.2.6 (April 9, 2017)
 - fixed upload progress bar